requirements.txt
langchain-community>=0.3.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.1.0
langchain-openai>=0.2.0
pydantic-settings>=2.0.0
langgraph>=0.2.0
chromadb>=0.5.0
pytest>=8.0.0
ragas>=0.4.0
datasets>=2.0.0
pyyaml>=6.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
sse-starlette>=2.0.0
mcp>=1.0.0,<2.0.0
